requirements.txt
beautifulsoup4>=4.9.3
requests>=2.25.1
pandas>=1.2.0
numpy>=1.19.2
pyyaml>=6.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
xgboost>=1.5.0
pytest>=7.0.0